+++ template/mock/data/b.js
@@ -3,7 +3,12 @@
  */
 module.exports = {
     '/api/api1': {
-        $params:['a2'],
+        $params:['a1'],
+        $method:'post',
+        $contentType:'application/json',
+        $specParams:{
+            a1:1
+        },
         $res:{
             data: 3
         }